fix: Wrong prices for multiple items
