+++ src/graphql/items.ts
@@ -38,30 +38,30 @@ export const GET_USER_OFFER_ITEMS = gql`
   ${ItemWithOfferFragment}
 `;
 
-export const GET_OFFER = addTypenameToDocument(
-  parse(`
-  query getOffer($id: ID!, $storeId: ID) {
-    offer(id: $id) {
-      ...Offer
-      item {
-        id
-        type
-        tokenId
-        contractAddress
-        supply
-        totalSupply
-        chainId
-      }
-      ... on MintOffer {
-        ...MintOffer
-      }
+export const GET_OFFER_DOCUMENT = `
+query getOffer($id: ID!, $storeId: ID) {
+  offer(id: $id) {
+    ...Offer
+    item {
+      id
+      type
+      tokenId
+      contractAddress
+      supply
+      totalSupply
+      chainId
+    }
+    ... on MintOffer {
+      ...MintOffer
     }
   }
+}
 
-  ${OfferFragment}
-  ${MintOfferFragment}
-`)
-);
+${OfferFragment}
+${MintOfferFragment}
+`;
+
+export const GET_OFFER = addTypenameToDocument(parse(GET_OFFER_DOCUMENT));
 
 export const GET_USER_ITEMS = gql`
   query getUserItems(
